+++ src/main/java/com/isroot/stash/plugin/YaccPreReceiveHook.java
@@ -21,6 +21,9 @@ import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
 
+import javax.annotation.Nonnull;
+import java.util.Collection;
+
 /**
  * @author Uldis Ansmits
  * @author Jim Bethancourt
@@ -56,7 +59,6 @@ public class YaccPreReceiveHook implements PreReceiveHook {
             }
         });
 
-        Settings settings = repositoryHookService.createSettingsBuilder().build(); // generate a default settings object
         if (hook.isEnabled() && hook.isConfigured()) {
             // Repository hook is configured and enabled.
             // Repository hook overrides default pre-receive hook configuration
@@ -65,36 +67,10 @@ public class YaccPreReceiveHook implements PreReceiveHook {
         } else {
             // Repository hook not configured
             log.debug("PreReceiveRepositoryHook not configured. Run PreReceiveHook");
-            PluginSettings pluginSettings = pluginSettingsFactory.createGlobalSettings();
 
-            Map<String, Object> settingsMap =
-                    (HashMap<String, Object>) pluginSettings.get(YaccConfigServlet.SETTINGS_MAP);
+            Settings storedConfig = YaccUtils.buildYaccConfig(pluginSettingsFactory, repositoryHookService);
 
-            Settings storedConfig = buildYaccConfig(settingsMap);
-
-            if (storedConfig != null) {
-                settings = storedConfig;
-            }
-        }
-
-        return yaccHook.onReceive(new RepositoryHookContext(repository, settings), refChanges, hookResponse);
-    }
-
-    Settings buildYaccConfig(Map<String, Object> settingsMap) {
-        HashMap<String, Object> config = new HashMap<String, Object>();
-        for (String fieldName : settingsMap.keySet()) {
-            addFieldValueToPluginConfigMap(settingsMap, config, fieldName);
-        }
-
-        return repositoryHookService.createSettingsBuilder().addAll(config).build();
-    }
-
-    void addFieldValueToPluginConfigMap(Map<String, Object> settingsMap, HashMap<String, Object> config, String fieldName) {
-        String value = (String) settingsMap.get(fieldName);
-        if (value != null && (value.equals("on") || value.equals("true"))) { // handle "on" value
-            config.put(fieldName, true);
-        } else if (value != null && !value.isEmpty()) {
-            config.put(fieldName, value);
+            return yaccHook.onReceive(new RepositoryHookContext(repository, storedConfig), refChanges, hookResponse);
         }
     }
 }