+++ src/test/java/it/com/isroot/stash/plugin/YaccGlobalSettingsPage.java
@@ -18,14 +18,7 @@ public class YaccGlobalSettingsPage extends YaccSettingsCommon {
 
     public YaccSettingsCommon clickSubmit() {
         submit.click();
-
-        // temp sleep
-        try {
-            Thread.sleep(5000);
-        } catch (InterruptedException e) {
-
-        }
-
+        waitABitForPageLoad();
         return this;
     }