+++ src/test/java/it/com/isroot/stash/plugin/YaccRepoSettingsPage.java
@@ -11,6 +11,9 @@ public class YaccRepoSettingsPage extends YaccSettingsCommon {
     @ElementBy(cssSelector = "tr[data-key=\"com.isroot.stash.plugin.yacc:yaccHook\"] .edit-button")
     private PageElement editYacc;
 
+    @ElementBy(cssSelector = "tr[data-key=\"com.isroot.stash.plugin.yacc:yaccHook\"] .mode-disabled")
+    private PageElement disableYacc;
+
     @ElementBy(className = "button-panel-submit-button")
     private PageElement submitButton;
 
@@ -26,14 +29,12 @@ public class YaccRepoSettingsPage extends YaccSettingsCommon {
 
     public YaccRepoSettingsPage clickSubmit() {
         submitButton.click();
+        waitABitForPageLoad();
+        return this;
+    }
 
-        // temp sleep
-        try {
-            Thread.sleep(5000);
-        } catch (InterruptedException e) {
-
-        }
-
+    public YaccRepoSettingsPage clickDisable() {
+        disableYacc.click();
         return this;
     }
 }