+++ src/test/java/ut/com/isroot/stash/plugin/YaccPreReceiveHookTest.java
@@ -8,13 +8,15 @@ import com.atlassian.bitbucket.repository.RefChange;
 import com.atlassian.bitbucket.repository.RefChangeType;
 import com.atlassian.bitbucket.repository.Repository;
 import com.atlassian.bitbucket.setting.Settings;
-import com.atlassian.bitbucket.setting.SettingsBuilder;
 import com.atlassian.bitbucket.user.EscalatedSecurityContext;
 import com.atlassian.bitbucket.user.SecurityService;
 import com.atlassian.bitbucket.util.UncheckedOperation;
 import com.atlassian.sal.api.pluginsettings.PluginSettings;
 import com.atlassian.sal.api.pluginsettings.PluginSettingsFactory;
+
+import com.atlassian.sal.testresources.pluginsettings.MockPluginSettingsFactory;
 import com.google.common.collect.Lists;
+import com.isroot.stash.plugin.YaccConfigServlet;
 import com.isroot.stash.plugin.YaccPreReceiveHook;
 import com.isroot.stash.plugin.YaccService;
 import com.isroot.stash.plugin.errors.YaccError;
@@ -23,6 +25,11 @@ import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
+import org.mockito.ArgumentCaptor;
+import org.mockito.Captor;
+import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
+import ut.com.isroot.stash.plugin.mock.MockSettingsBuilder;
 
 import java.io.PrintWriter;
 import java.io.StringWriter;
@@ -32,37 +39,49 @@ import java.util.List;
 import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.data.MapEntry.entry;
 import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.anyMap;
 import static org.mockito.Matchers.anyString;
 import static org.mockito.Mockito.*;
 
+
 /**
- * @author Jim Bethancourt
- *
  * Cloned from YaccHookTest.java and modified to test against the PreReceiveHook interface.
+ *
+ * @author Jim Bethancourt
  */
 public class YaccPreReceiveHookTest {
-    @Mock private YaccService yaccService;
-    @Mock private HookResponse hookResponse;
-    @Mock private RepositoryHookService repositoryHookService;
-    @Mock private RepositoryHook repositoryHook;
-    @Mock private Repository repository;
-    @Mock private SecurityService securityService;
-    @Mock private EscalatedSecurityContext escalatedSecurityContext;
-    @Mock private PluginSettingsFactory pluginSettingsFactory;
-    @Mock private PluginSettings pluginSettings;
-    @Mock private SettingsBuilder settingsBuilder;
-    @Mock private Settings settings;
-    private Map<String, Object> settingsMap = new HashMap<String, Object>();
+    @Mock
+    private YaccService yaccService;
+    @Mock
+    private HookResponse hookResponse;
+    @Mock
+    private RepositoryHookService repositoryHookService;
+    @Mock
+    private RepositoryHook repositoryHook;
+    @Mock
+    private Repository repository;
+    @Mock
+    private SecurityService securityService;
+    @Mock
+    private EscalatedSecurityContext escalatedSecurityContext;
+
+    @Captor
+    private ArgumentCaptor<Settings> settingsCapture;
+
+    private PluginSettingsFactory pluginSettingsFactory;
+
+    private Map<String, Object> globalSettingsMap = new HashMap<>();
 
     private StringWriter errorMessage;
     private YaccPreReceiveHook yaccPreReceiveHook;
 
-
     @Before
     public void setup() throws Throwable {
         MockitoAnnotations.initMocks(this);
+
+        pluginSettingsFactory = new MockPluginSettingsFactory();
 
         yaccPreReceiveHook = new YaccPreReceiveHook(yaccService,
                 pluginSettingsFactory, securityService, repositoryHookService);
@@ -72,20 +91,17 @@ public class YaccPreReceiveHookTest {
                 .thenReturn(escalatedSecurityContext);
         when(escalatedSecurityContext.call(any(UncheckedOperation.class))).thenReturn(repositoryHook);
 
-        when(repositoryHookService.createSettingsBuilder()).thenReturn(settingsBuilder);
-        when(settingsBuilder.build()).thenReturn(settings);
-        when(settingsBuilder.addAll(anyMap())).thenReturn(settingsBuilder);
-
-        when(pluginSettingsFactory.createGlobalSettings()).thenReturn(pluginSettings);
-        when(pluginSettings.get(anyString())).thenReturn(settingsMap);
+        when(repositoryHookService.createSettingsBuilder()).thenReturn(new MockSettingsBuilder());
 
         errorMessage = new StringWriter();
         when(hookResponse.err()).thenReturn(new PrintWriter(errorMessage));
 
+        PluginSettings pluginSettings = pluginSettingsFactory.createGlobalSettings();
+        pluginSettings.put(YaccConfigServlet.SETTINGS_MAP, globalSettingsMap);
     }
 
     @Test
-    public void testOnReceive_repositoryHookConfigured(){
+    public void testOnReceive_repositoryHookConfigured() {
         when(repositoryHook.isConfigured()).thenReturn(true);
         when(repositoryHook.isEnabled()).thenReturn(true);
 
@@ -145,7 +161,7 @@ public class YaccPreReceiveHookTest {
         when(yaccService.checkRefChange(any(Repository.class), any(Settings.class), any(RefChange.class)))
                 .thenReturn(Lists.newArrayList(new YaccError("error1")));
 
-        when(settings.getString("errorMessageHeader")).thenReturn("");
+        globalSettingsMap.put("errorMessageHeader", "");
 
         yaccPreReceiveHook.onReceive(repository, getMockRefChanges(), hookResponse);
 
@@ -157,7 +173,7 @@ public class YaccPreReceiveHookTest {
         when(yaccService.checkRefChange(any(Repository.class), any(Settings.class), any(RefChange.class)))
                 .thenReturn(Lists.newArrayList(new YaccError("error1")));
 
-        when(settings.getString("errorMessageHeader")).thenReturn("Custom Header");
+        globalSettingsMap.put("errorMessageHeader", "Custom Header");
 
         yaccPreReceiveHook.onReceive(repository, getMockRefChanges(), hookResponse);
 
@@ -171,11 +187,37 @@ public class YaccPreReceiveHookTest {
         when(yaccService.checkRefChange(any(Repository.class), any(Settings.class), any(RefChange.class)))
                 .thenReturn(Lists.newArrayList(new YaccError("error1")));
 
-        when(settings.getString("errorMessageFooter")).thenReturn("Custom Footer");
+        globalSettingsMap.put("errorMessageFooter", "Custom Footer");
 
         yaccPreReceiveHook.onReceive(repository, getMockRefChanges(), hookResponse);
 
         assertThat(errorMessage.toString()).endsWith("\nCustom Footer\n\n");
+    }
+
+    @Test
+    public void testOnReceive_nullSettingsMap_hookWorksBeforeItHasBeenConfigured() {
+        pluginSettingsFactory.createGlobalSettings().put(YaccConfigServlet.SETTINGS_MAP, null);
+
+        boolean isPushAllowed = yaccPreReceiveHook.onReceive(repository, getMockRefChanges(), hookResponse);
+
+        assertThat(isPushAllowed).isTrue();
+    }
+
+    @Test
+    public void testOnReceive_globalHookSettingsPassedToHook() {
+        globalSettingsMap.put("commitMessageRegex", "bar");
+        globalSettingsMap.put("requireMatchingAuthorEmail", "true");
+
+        yaccPreReceiveHook.onReceive(repository, getMockRefChanges(), hookResponse);
+
+        verify(yaccService).checkRefChange(eq(repository), settingsCapture.capture(), any(RefChange.class));
+
+        Settings hookSettings = settingsCapture.getValue();
+
+        assertThat(hookSettings.asMap())
+                .contains(
+                        entry("commitMessageRegex", "bar"),
+                        entry("requireMatchingAuthorEmail", true));
     }
 
     private List<RefChange> getMockRefChanges() {